realCaller/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Faker==20.1.0
Python==3.11.5